+++ src/main/java/com/farcr/nomansland/common/mixin/client/MinecraftMixin.java
@@ -6,14 +6,12 @@ import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import net.minecraft.client.DeltaTracker;
 import com.farcr.nomansland.client.renderer.dreams.ClientDreamRenderer;
-import com.farcr.nomansland.common.dreams.DreamManager;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.*;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import net.minecraft.client.sounds.SoundManager;
 import org.spongepowered.asm.mixin.Final;
-import net.minecraft.client.gui.screens.InBedChatScreen;
 import net.minecraft.client.gui.screens.Screen;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -62,10 +60,10 @@ public abstract class MinecraftMixin {
     }
 
     @WrapOperation(method = "updateScreenAndTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/sounds/SoundManager;stop()V"))
-    private void nml$keepBellSoundThroughTransition(SoundManager soundManager, Operation<Void> original) {
-        if (InvertedBellClientHandler.instance.isActive()) {
-            return;
-        }
+    private void nml$keepSoundsThroughTransitions(SoundManager soundManager, Operation<Void> original) {
+        if (InvertedBellClientHandler.instance.isActive()) return;
+        ClientDreamRenderer renderer = ClientDreamRenderer.getInstance();
+        if (renderer.clientIsDreaming()) return;
         original.call(soundManager);
     }
 }